move structures out of webserver
