+++ client/ayon_core/tools/tray/webserver/__init__.py
@@ -1,10 +1,8 @@
-from .structures import HostMsgAction
 from .base_routes import RestApiEndpoint
 from .server import find_free_port, WebServerManager
 
 
 __all__ = (
-    "HostMsgAction",
     "RestApiEndpoint",
     "find_free_port",
     "WebServerManager",
